requirements.txt
python-telegram-bot==21.6
openai>=1.40
PyMuPDF>=1.24
python-docx>=1.1
youtube-transcript-api>=0.6
